+++ feature_forecast/src/main/java/com/example/feature_forecast/domain/WeatherInfo.kt
@@ -2,15 +2,15 @@ package com.example.feature_forecast.domain
 
 data class WeatherInfo(
     val cityName: String,
-    val day: Int,
+    val day: Long,
     val weatherDescription: String,
     val iconId: String,
     val temperatureMorning: Double,
     val temperatureDay: Double,
     val temperatureEvening: Double,
-    val humidity: Int,
+    val humidity: Long,
     val windSpeed: Double,
-    val pressure: Int,
-    val sunrise: Int,
-    val sunset: Int
+    val pressure: Long,
+    val sunrise: Long,
+    val sunset: Long
 )
